Next version pre features
- Improved popup rendering performance and quality
- Updated main UI elements with shadows
- Updated "About" page
- Fixed error when restoring a missing config

+++ src/main/java/de/pt400c/defaultsettings/FileUtil.java
@@ -21,7 +21,6 @@ import java.util.ArrayList;
 import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
-import java.util.TimeZone;
 import java.util.UUID;
 import java.util.regex.Pattern;
 import javax.xml.bind.DatatypeConverter;
@@ -137,16 +136,9 @@ public static void switchState(Byte state, String query) {
 
 			try (Reader reader = new FileReader(main)) {
 				persistentJson = gson.fromJson(reader, PersistentJSON.class);
-
 			} catch (Exception e) {
-				DefaultSettings.log.log(Level.ERROR, "Exception at processing persistent configs: ", e);
-				persistentJson = new PersistentJSON();
-				persistentJson.save(new File(mcDataDir, persistentLocation));
-			}
 
-		} else {
-			persistentJson = new PersistentJSON();
-			persistentJson.save(new File(mcDataDir, persistentLocation));
+			}
 		}
 
 		return persistentJson;
@@ -197,9 +189,41 @@ public static void switchState(Byte state, String query) {
 		final String identifier = mainJson.getIdentifier();
 		
 		if(!getIdentifier().equals(identifier))
-			mainJson.setIdentifier(identifier);
+			mainJson.setIdentifier(getIdentifier());
+
+		File persFile = new File(mcDataDir, persistentLocation);
+		if (persFile.exists()) {
+			getPersistent().check.forEach((k, v) -> mainJson.check.put(k, v));
+			persFile.delete();
+		}
 		
 		mainJson.save(main);
+	}
+	
+	public static void initUUID() throws NoSuchAlgorithmException {
+		DefaultSettings.PLAYER_UUID = MC.getSession().getProfile().getId().toString();
+		final File main = new File(mcDataDir, mainLocation);
+		if (getMainJSON().created_for == null || getMainJSON().created_for.equals("NEW")) {
+
+			try {
+				mainJson.created_for = getUUID(DefaultSettings.PLAYER_UUID);
+			} catch (NoSuchAlgorithmException e) {
+				e.printStackTrace();
+			}
+
+			mainJson.save(main);
+		} else {
+
+			final String created_for = mainJson.created_for;
+
+			if (!getUUID(DefaultSettings.PLAYER_UUID).equals(created_for)) {
+				mainJson.created_for = getUUID(DefaultSettings.PLAYER_UUID);
+				mainJson.check.clear();
+
+			}
+			
+			mainJson.save(main);
+		}
 	}
 	
 	/**
@@ -235,7 +259,8 @@ public static void switchState(Byte state, String query) {
 			} catch (UnknownHostException | SocketException | NoSuchAlgorithmException e) {
 				DefaultSettings.log.log(Level.ERROR, "Exception at processing configs: ", e);
 			}
-			mainJson = new MainJSON().setVersion(DefaultSettings.VERSION).setIdentifier(identifier).setCreated(formatter.format(date) + " (" + TimeZone.getDefault().getDisplayName() + ")");
+			mainJson = new MainJSON().setVersion(DefaultSettings.VERSION).setIdentifier(identifier).setCreated(formatter.format(date));
+			mainJson.created_for = "NEW";
 			mainJson.initPopup = true;
 			File fileDir = new File(mcDataDir, "config");
 			for (File file : fileDir.listFiles(fileFilter)) 
@@ -280,7 +305,7 @@ public static void switchState(Byte state, String query) {
 			mainJson.save(main);
 		}else {
 			for(String name : getOverrides().keySet()) 
-				if(getActives().contains(name) && (!getPersistent().check.containsKey(name) || !getPersistent().check.get(name).equals(mainJson.overrides.get(name))))
+				if(getActives().contains(name) && (!getMainJSON().check.containsKey(name) || !getMainJSON().check.get(name).equals(mainJson.overrides.get(name)))) 
 						restoreSingleConfig(name);
 				
 			final File main = new File(mcDataDir, mainLocation);
@@ -322,16 +347,21 @@ public static void switchState(Byte state, String query) {
 	public static void restoreSingleConfig(String name) throws IOException {
 		try {
 			File file = new File(getMainFolder(), name);
-			if(file.isDirectory()) 
-				FileUtils.copyDirectory(file, new File(mcDataDir, "config/" + name));
-			
-			else 
-				FileUtils.copyFile(file, new File(mcDataDir, "config/" + name));
+			if (file.exists()) {
+				if (file.isDirectory())
+					FileUtils.copyDirectory(file, new File(mcDataDir, "config/" + name));
+
+				else
+					FileUtils.copyFile(file, new File(mcDataDir, "config/" + name));
+			}else {
+				DefaultSettings.log.log(Level.WARN, "Couldn't restore a config file as it's missing: " + name);
+				return;
+			}
 			
 			String random = getOverrides().get(name);
 			
-			getPersistent().check.put(name, random);
-			persistentJson.save(new File(mcDataDir, persistentLocation));
+			getMainJSON().check.put(name, random);
+			mainJson.save(new File(mcDataDir, mainLocation));
 
 		} catch (IOException e) {
 			throw e;
@@ -534,7 +564,11 @@ public static void switchState(Byte state, String query) {
 
 	public static void restoreServers() throws IOException {
 		try {
-			FileUtils.copyFile(new File(getMainFolder(), "servers.dat"), new File(mcDataDir, "servers.dat"));
+			File file = new File(getMainFolder(), "servers.dat");
+			if(file.exists())
+				FileUtils.copyFile(file, new File(mcDataDir, "servers.dat"));
+			else
+				DefaultSettings.log.log(Level.WARN, "Couldn't restore the server config as it's not included");
 		} catch (IOException e) {
 			DefaultSettings.log.log(Level.ERROR, "Couldn't restore the server config: ", e);
 		}
@@ -554,6 +588,15 @@ public static void switchState(Byte state, String query) {
 		} finally {
 			writer.close();
 		}
+	}
+	
+	public static String getUUID(String uuid) throws NoSuchAlgorithmException {
+
+		MessageDigest md = MessageDigest.getInstance("MD5");
+		md.update(uuid.getBytes());
+		byte[] digest = md.digest();
+		return DatatypeConverter.printHexBinary(digest).toUpperCase();
+
 	}
 
 	public static void saveOptions() throws NullPointerException, IOException {
